requirements.txt
attrs==19.3.0
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.4.3
pytest-mock==3.1.1
PyYAML==5.3.1
six==1.15.0
wcwidth==0.2.5
